Salda deuda tecnica: campo muerto, plata en float, adaptadores duplicados
- Elimina declared_value_ars (se pedia, se validaba, nunca se leia) y el
  get_db() sin uso.
- Decimal + ROUND_HALF_UP para toda la plata, en vez de float/round().
- HttpCarrierAdapter generico por composicion: los 3 transportistas dejan
  de repetir el mismo try/except/timeout y quedan como endpoint + 2
  funciones de mapeo.
- SQLiteQuoteHistory usa asyncio.to_thread para no bloquear el event loop.
- DATABASE_URL configurable por env var.
- Alembic para migraciones (standalone, no enganchado al lifespan a
  proposito: los tests parchean el engine, no la URL).
- logging.basicConfig + warning en cada falla de carrier.
- CI: pytest en cada push via GitHub Actions.

+++ app/ports/carrier_port.py
@@ -1,5 +1,6 @@
 from abc import ABC, abstractmethod
 from dataclasses import dataclass
+from decimal import Decimal
 
 from app.domain.package import Package
 from app.domain.trace import Tracer
@@ -18,7 +19,7 @@ CARRIER_REQUEST_TIMEOUT_S = 2.0
 
 @dataclass(frozen=True)
 class CarrierQuote:
-    amount_ars: float
+    amount_ars: Decimal
     eta_days: int
